Updating all of DREAM3D for IGEometry API Changes.
Updates #262

+++ Source/DREAM3DLib/Geometry/MeshStructs.h
@@ -62,8 +62,8 @@ namespace SurfaceMesh
 
     typedef struct
     {
-      int site_id[4];    // stores 4 sites at the corners of each square...
-      int edge_id[4];    // stores edge id turned on...others will have dummy -1...
+      int64_t site_id[4];    // stores 4 sites at the corners of each square...
+      int64_t edge_id[4];    // stores edge id turned on...others will have dummy -1...
       int nEdge;         // number of edges on the square...
       int FCnode;        // face center node...if not, it's -1...
       int effect;        // 0 if the square is useless; 1 is good...
@@ -71,7 +71,7 @@ namespace SurfaceMesh
 
     typedef struct
     {
-      int node_id[2];    // the segment heads from node_id[0] to node_id[1]...
+      int64_t node_id[2];    // the segment heads from node_id[0] to node_id[1]...
       int edgeKind;      // initially marked as 2; for face edges it's always 2...
       int nSpin[2];      // 0 is to the left of the arrow; 1 is at right...
     } Segment;
@@ -79,14 +79,14 @@ namespace SurfaceMesh
     /* Used for "inner edge" spin calculations */
     typedef struct
     {
-      int node_id[2];    // the segment heads from node_id[0] to node_id[1]...
+      int64_t node_id[2];    // the segment heads from node_id[0] to node_id[1]...
       int edgeKind;      // initially marked with 2...
       int nSpin[4];
     } ISegment;
 
     typedef struct
     {
-      int node_id[3];    // stores three new node id for vertices of the triangles...
+      int64_t node_id[3];    // stores three new node id for vertices of the triangles...
       uint64_t e_id[3];       // stores three new edge id for sides of the triangles...
       int nSpin[2];      // two spins...
       int edgePlace[3];
@@ -107,7 +107,7 @@ namespace SurfaceMesh
   {
     typedef struct
     {
-      int nodeId;
+      int64_t nodeId;
       int nodeKind;
       float x;
       float y;
@@ -120,10 +120,10 @@ namespace SurfaceMesh
   {
     typedef struct
     {
-      int triId;
-      int nodeId_0;
-      int nodeId_1;
-      int nodeId_2;
+      int64_t triId;
+      int64_t nodeId_0;
+      int64_t nodeId_1;
+      int64_t nodeId_2;
       int label_0;
       int label_1;
     } TrianglesFileRecord_t;
